+++ veamy/src/physics/materials/Material.cpp
@@ -34,7 +34,7 @@ bool Material::operator==(const Material &other) const{
             std::abs(this->material_v-other.material_v) < config->getTolerance();
 }
 
-void Material::setMultiplicativeFactor(int f) {
+void Material::setMultiplicativeFactor(double f) {
     this->factor = f;
 }
